user metrics: add licensing, non-licensing segments

+++ test/unit/numbers/numbers_csv_test.rb
@@ -1,5 +1,6 @@
 require 'test_helper'
 require 'numbers/numbers_csv'
+require 'numbers/user_segments'
 
 class NumbersCsvTest < ActiveSupport::TestCase
   include FactoryGirl::Syntax::Methods
@@ -10,7 +11,7 @@ class NumbersCsvTest < ActiveSupport::TestCase
   end
 
   def teardown
-    `rm ./numbers.csv`
+    `rm ./numbers.csv` if File.exist?("./numbers.csv")
   end
 
   def numbers_csv
@@ -105,5 +106,26 @@ class NumbersCsvTest < ActiveSupport::TestCase
     NumbersCsv.generate
     assert numbers_csv.include? ["Active accounts count by email domain", "admin.example.com", "1"]
     assert numbers_csv.include? ["Active accounts count by email domain", "example.com", "3"]
+  end
+
+  test "licensing segmenting" do
+    licensing = create(:application, name: "Licensing")
+    create(:supported_permission, name: "signin", application: licensing)
+    another_app = create(:application, name: "Another app")
+    create(:supported_permission, name: "signin", application: another_app)
+
+    users = create_list(:user, 4)
+    users[0].grant_permission(licensing, "signin")
+
+    users[1].grant_permission(licensing, "signin")
+    users[1].grant_permission(another_app, "signin")
+
+    users[2].grant_permission(licensing, "another perm")
+
+    users.each { |u| u.extend(UserSegments::UserSegmentExtensions) }
+
+    assert users[0].licensing_user?
+    assert !users[1].licensing_user?
+    assert !users[2].licensing_user?
   end
 end
